Changed a lot of stuff mostly better understanding of data layouts and such. Multithreading still really good maybe better model loading. Really want to impliment physics rotations and then animations afterwards but this is all Ive gotten done.

+++ hw3d/Threads.h
@@ -17,7 +17,7 @@ private:
 
 		void exeWork();
 
-		std::atomic<unsigned int> lWaiting = 0;
+		std::atomic<uint8_t> lWaiting = 0;
 		void checkWork(unsigned int uWid);
 		std::mutex wCountMTX;
 		std::vector<uint8_t> Queue;
